requirements.txt
# 汽车电源测试框架依赖
# 使用ruamel.yaml替代pyyaml，避免Windows编译问题

# 测试框架核心
pytest==7.4.0
pytest-xdist==3.3.0
pytest-html==4.1.0
allure-pytest==2.13.0

# 硬件通信与控制
pyvisa==1.13.0
python-can==4.2.0

# 数据配置处理（使用ruamel.yaml替代pyyaml）
ruamel.yaml==0.17.21

# 基础工具
colorama==0.4.6
loguru==0.7.2
